+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 * `player` configuration option is now a list.
+* File pattern now matches only against the base filename.
 
 ## 0.13.0 - 2018-08-20